[SWM-323] Refactor : replace material service to version 2
